@@ -1,6 +1,6 @@
 {
-  "name": "movement-tracker",
-  "version": "4.0.1",
+  "name": "app",
+  "version": "4.0.4",
   "private": true,
   "description": "An app for tracking Christian discipleship trees",
   "author": "Nick Ellis <ellis.nick96@gmail.com>",
@@ -20,43 +20,50 @@
     "test:functions:emulated": "firebase emulators:exec --only auth,functions,firestore,database \"jest --watchAll\""
   },
   "dependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.16.8",
-    "@babel/preset-env": "^7.16.11",
+    "@babel/plugin-transform-modules-commonjs": "^7.18.2",
+    "@babel/preset-env": "^7.18.2",
     "@firebase/rules-unit-testing": "^2.0.2",
-    "@quasar/extras": "^1.12.5",
-    "babel-jest": "^27.5.0",
-    "chart.js": "^3.6.0",
-    "core-js": "^3.21.0",
+    "@quasar/extras": "^1.14.0",
+    "babel-jest": "^28.1.0",
+    "chart.js": "^3.8.0",
+    "core-js": "^3.22.7",
     "esm": "^3.2.25",
     "file-saver": "^2.0.5",
-    "firebase": "^9.6.6",
-    "firebase-functions": "^3.17.1",
-    "quasar": "^2.5.3",
-    "ts-jest": "^27.1.3",
-    "vue": "^3.2.29",
-    "vue-router": "^4.0.12",
+    "firebase": "^9.8.1",
+    "quasar": "^2.7.1",
+    "ts-jest": "^27.1.5",
+    "vue": "^3.2.36",
+    "vue-chart-3": "^3.0.9",
+    "vue-router": "^4.0.15",
     "vuedraggable": "^4.1.0",
     "vuex": "^4.0.1"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.17.0",
-    "@quasar/app": "^3.3.2",
+    "@babel/eslint-parser": "^7.18.2",
+    "@quasar/app": "^3.3.3",
     "@quasar/quasar-app-extension-qcalendar": "^4.0.0-beta.3",
-    "@quasar/quasar-app-extension-testing-e2e-cypress": "^4.0.1",
-    "@types/node": "^17.0.15",
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
+    "@quasar/quasar-app-extension-testing-e2e-cypress": "^4.1.2",
+    "@types/node": "^17.0.35",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
     "cross-env": "^7.0.3",
-    "cypress": "^9.4.1",
+    "cypress": "^9.7.0",
     "cypress-firebase": "^2.2.1",
-    "eslint": "^8.8.0",
+    "eslint": "^8.16.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-vue": "^8.4.1",
-    "firebase-admin": "^10.0.2",
-    "stylus": "^0.56.0",
+    "firebase-admin": "^10.2.0",
+    "js-base64": "^3.7.2",
+    "stylus": "^0.57.0",
     "stylus-loader": "^6.1.0",
     "workbox-webpack-plugin": "^6.2.2"
+  },
+  "overrides": {
+    "async@<3.2.2": "3.2.2",
+    "minimist@1.2.5": "1.2.6",
+    "node-ipc@>9.2.1 <10": "9.2.1",
+    "node-ipc@>10.1.0": "10.1.0"
   },
   "browserslist": [
     "last 10 Chrome versions",
@@ -70,8 +77,8 @@
     "last 5 Opera versions"
   ],
   "engines": {
-    "node": ">= 12.22.1",
-    "npm": ">= 6.13.4",
+    "node": ">=16.14.2",
+    "npm": ">=8.7.0",
     "yarn": ">= 1.21.1"
   }
 }
