@@ -11,26 +11,37 @@
     "logs": "firebase functions:log",
     "test": "jest --watchAll --verbose=false",
     "test:converter": "firebase emulators:exec --only firestore,database --import=./../server_data \"jest converter._jest.spec.js\"",
-    "emulateSnapshot": "firebase emulators:start --only firestore,database --import=./../server_data "
+    "emulateSnapshot": "firebase emulators:start --only firestore,database,auth,functions --import=./../server_data "
   },
   "main": "lib/index.js",
   "dependencies": {
-    "@google-cloud/pubsub": "^2.18.5",
-    "firebase-admin": "^10.0.2",
-    "firebase-functions": "^3.17.1"
+    "@google-cloud/pubsub": "^2.19.4",
+    "firebase-admin": "^10.2.0",
+    "firebase-functions": "^3.21.2",
+    "firebase-tools": "^11.0.1",
+    "vm2": "^3.9.9"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.10.2",
-    "@typescript-eslint/parser": "^5.10.2",
-    "eslint": "^8.8.0",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
+    "eslint": "^8.16.0",
     "eslint-config-google": "^0.14.0",
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-promise": "^6.0.0",
-    "firebase-functions-test": "^0.3.1",
-    "typescript": "^4.5.5"
+    "firebase-functions-test": "^2.2.0",
+    "typescript": "^4.7.2"
+  },
+  "overrides": {
+    "async@<3.2.2": "3.2.2",
+    "minimist@<1.2.6": "1.2.6",
+    "node-forge@<1.3.0": "1.3.0",
+    "node-ipc@>9.2.1 <10": "9.2.1",
+    "node-ipc@>10.1.0": "10.1.0"
   },
   "private": true,
   "engines": {
-    "node": "16"
-  }
+    "node": "16",
+    "npm": ">=8.7.0"
+  },
+  "version": "4.0.3"
 }
