@@ -1,32 +1,40 @@
 {
   "name": "movement-tracker",
-  "version": "4.0.0",
+  "version": "4.0.4",
   "private": true,
   "description": "An app for tracking Christian discipleship trees",
   "author": "Nick Ellis <ellis.nick96@gmail.com>",
   "productName": "Movement Tracker",
   "scripts": {
     "start:emulators": "firebase emulators:start --only auth,functions,firestore,database",
-    "start:emulators-import": "firebase emulators:start --only auth,functions,firestore,pubsub --import ./firestoreData",
+    "start:emulators-import": "firebase emulators:start --only auth,functions,firestore,pubsub --import ./server_data",
     "test:rules:fire": "firebase emulators:exec --only firestore \"jest firestore.rules._jest.spec.js --detectOpenHandles\"",
     "test:rules:rtdb": "firebase emulators:exec --only database \"jest database.rules._jest.spec.js --detectOpenHandles\"",
     "test:rules:fire:watch": "firebase emulators:exec --only firestore \"jest firestore.rules._jest.spec.js --detectOpenHandles --watchAll\"",
     "test:rules:rtdb:watch": "firebase emulators:exec --only database \"jest database.rules._jest.spec.js --detectOpenHandles --watchAll\""
   },
   "dependencies": {
-    "firebase": "^9.6.5"
+    "firebase": "^9.6.11",
+    "vm2": "^3.9.9"
   },
   "devDependencies": {
     "@firebase/rules-unit-testing": "^2.0.2",
     "@quasar/babel-preset-app": "^2.0.1",
-    "babel-jest": "^27.4.6",
+    "babel-jest": "^27.5.1",
     "fs-extra": "^10.0.0",
     "jest": "^27.4.7",
     "ts-jest": "^27.1.3"
   },
+  "overrides": {
+    "async@<3.2.2": "3.2.2",
+    "minimist@<1.2.6": "1.2.6",
+    "node-forge@<1.3.0": "1.3.0",
+    "node-ipc@>9.2.1 <10": "9.2.1",
+    "node-ipc@>10.1.0": "10.1.0"
+  },
   "engines": {
-    "node": ">= 12.22.1",
-    "npm": ">= 6.13.4",
+    "node": ">=16.14.2",
+    "npm": ">=8.7.0",
     "yarn": ">= 1.21.1"
   }
 }
