@@ -28,6 +28,7 @@
           v-if="props.row.display"
           @contextmenu.stop
           data-cy="member"
+          :data-cy-id="props.row.id"
         >
           <q-tooltip
             v-if="props.row.notes > ''"
@@ -65,15 +66,15 @@
               :class="{ 'text-underline': props.row.display.underline }"
               :style="
                 'background-color:' +
-                props.row.display.background +
-                '; color:' +
-                props.row.display.color +
-                '; ' +
-                'font-size:' +
-                size +
-                '%;border-color:' +
-                props.row.display.outline +
-                ' !important;'
+                  props.row.display.background +
+                  '; color:' +
+                  props.row.display.color +
+                  '; ' +
+                  'font-size:' +
+                  size +
+                  '%;border-color:' +
+                  props.row.display.outline +
+                  ' !important;'
               "
               style="border-width: 3px; border-style: solid; z-index: 10"
               :rounded="props.row.display.round"
@@ -97,7 +98,7 @@
               ? props.row.role.label
               : roles[props.row.role]
               ? roles[props.row.role].label
-              : "--Missing Role--"
+              : '--Missing Role--'
           }}</q-td>
           <q-td key="mods" :props="props" class="q-gutter-sm" data-cy="mods">
             <div class="text-left q-gutter-xs">
@@ -124,28 +125,22 @@
 </template>
 
 <script>
-import { useStore } from "vuex";
-import { defineAsyncComponent, ref, onMounted, computed, watch } from "vue";
-const memberJS = require("./../scripts/member.js");
-import { LocalStorage } from "quasar";
-import { useRoute } from "vue-router";
+import { useStore } from 'vuex';
+import { defineAsyncComponent, ref, onMounted, computed, watch } from 'vue';
+const memberJS = require('./../scripts/member.js');
+import { useQuasar } from 'quasar';
+import { useRoute } from 'vue-router';
 // const promiseDebounce = require('./../scripts/promiseDebounce.js')
 
 export default {
-  props: ["readOnly", "members", "mods", "roles"],
+  props: ['readOnly', 'members', 'mods', 'roles'],
   setup(props) {
+    const q = useQuasar();
     const store = useStore();
     const route = useRoute();
     const columns = ref([]);
     const rows = ref([]);
-    const pagination = ref({
-      sortBy: "name",
-      descending: true,
-      page: 1,
-      rowsPerPage: 10,
-      rowsNumber: 10,
-    });
-    const filter = ref("");
+    const filter = ref('');
     const loading = ref(false);
     const membersSelected = ref([]);
     const isFullscreen = ref(false);
@@ -156,85 +151,99 @@ export default {
     columns.value = permissions.members_edit
       ? [
           {
-            name: "preview",
-            align: "left",
-            label: "Preview",
-            field: "preview",
+            name: 'preview',
+            align: 'left',
+            label: 'Preview',
+            field: 'preview'
           },
           {
-            name: "name",
-            align: "left",
-            label: "Name",
-            field: "name",
-            sortable: true,
+            name: 'name',
+            align: 'left',
+            label: 'Name',
+            field: 'name',
+            sortable: true
           },
           {
-            name: "role",
-            align: "center",
-            label: "Role",
-            field: "role",
-            sortable: true,
+            name: 'role',
+            align: 'center',
+            label: 'Role',
+            field: 'role',
+            sortable: true
           },
           {
-            name: "mods",
-            align: "center",
-            label: "Modifiers",
-            field: "mods",
-            sortable: true,
+            name: 'mods',
+            align: 'center',
+            label: 'Modifiers',
+            field: 'mods',
+            sortable: true
           },
           {
-            name: "actions",
-            label: "Actions",
-            field: "actions",
-            align: "right",
-          },
+            name: 'actions',
+            label: 'Actions',
+            field: 'actions',
+            align: 'right'
+          }
         ]
       : [
           {
-            name: "preview",
-            align: "left",
-            label: "Preview",
-            field: "preview",
+            name: 'preview',
+            align: 'left',
+            label: 'Preview',
+            field: 'preview'
           },
           {
-            name: "name",
-            align: "left",
-            label: "Name",
-            field: "name",
-            sortable: true,
+            name: 'name',
+            align: 'left',
+            label: 'Name',
+            field: 'name',
+            sortable: true
           },
           {
-            name: "role",
-            align: "center",
-            label: "Role",
-            field: "role",
-            sortable: true,
+            name: 'role',
+            align: 'center',
+            label: 'Role',
+            field: 'role',
+            sortable: true
           },
           {
-            name: "mods",
-            align: "center",
-            label: "Modifiers",
-            field: "mods",
-            sortable: true,
-          },
+            name: 'mods',
+            align: 'center',
+            label: 'Modifiers',
+            field: 'mods',
+            sortable: true
+          }
         ];
-    pagination.value = LocalStorage.has("memberListPagination")
-      ? LocalStorage.getItem("memberListPagination")
+    const pagination = ref({
+      sortBy: 'name',
+      descending: false,
+      page: 1,
+      rowsPerPage: 10
+      // rowsNumber: xx if getting data from a server
+    });
+
+    pagination.value = q.localStorage.has('memberListPagination')
+      ? q.localStorage.getItem('memberListPagination')
       : {
-          sortBy: "name",
-          descending: true,
+          sortBy: 'name',
+          descending: false,
           page: 1,
-          rowsPerPage: 10,
-          rowsNumber: 10,
+          rowsPerPage: 10
+          // rowsNumber: xx if getting data from a server
         };
+    watch(pagination, (newVal, oldVal) => {
+      if (JSON.stringify(newVal) !== JSON.stringify(oldVal)) {
+        console.log(pagination.value);
+        q.localStorage.set('memberListPagination', pagination.value);
+      }
+    });
     function setFilterQuery(val) {
-      store.commit("movement/setFilterQuery", val);
+      store.commit('movement/setFilterQuery', val);
     }
 
     function membersFilterMethod(rows, terms, cols, cellValue) {
-      const lowerTerms = terms ? terms.toLowerCase() : "";
-      return rows.filter((row) => {
-        if (row.id === "root") return false;
+      const lowerTerms = terms ? terms.toLowerCase() : '';
+      return rows.filter(row => {
+        if (row.id === 'root') return false;
         else if (row.name.toLowerCase().indexOf(lowerTerms) !== -1) {
           return true;
         } else if (
@@ -245,7 +254,7 @@ export default {
         ) {
           return true;
         } else if (
-          row.mods.some((style) =>
+          row.mods.some(style =>
             props.mods[style]
               ? props.mods[style].label.toLowerCase().indexOf(lowerTerms) !== -1
               : false
@@ -256,13 +265,10 @@ export default {
         return false;
       });
     }
-    watch(pagination, () => {
-      LocalStorage.set("memberListPagination", pagination.value);
-    });
     watch(membersSelected, () => {
       membersSelectedFiltered.value = Array.from(
         membersSelected.value,
-        (x) => x.id
+        x => x.id
       );
     });
 
@@ -270,14 +276,12 @@ export default {
     // SELECT * FROM ... WHERE...LIMIT...
     function fetchFromServer(startRow, count, filter, sortBy, descending) {
       const data = filter
-        ? Object.values(props.members).filter((row) =>
-            row.name.includes(filter)
-          )
+        ? Object.values(props.members).filter(row => row.name.includes(filter))
         : Object.values(props.members).slice();
       // handle sortBy
       if (sortBy) {
         const sortFn =
-          sortBy === "name"
+          sortBy === 'name'
             ? descending
               ? (a, b) => (a.name > b.name ? -1 : a.name < b.name ? 1 : 0)
               : (a, b) => (a.name > b.name ? 1 : a.name < b.name ? -1 : 0)
@@ -295,7 +299,7 @@ export default {
         return Object.values(props.members).length;
       }
       let count = 0;
-      Object.values(props.members).forEach((treat) => {
+      Object.values(props.members).forEach(treat => {
         if (treat.name.includes(filter)) {
           ++count;
         }
@@ -343,13 +347,13 @@ export default {
 
     onMounted(() => {
       if (route.params.snapId) {
-        store.dispatch("snapshot/fetchMembers", {
+        store.dispatch('snapshot/fetchMembers', {
           movId: route.params.movId,
-          snapId: route.params.snapId,
+          snapId: route.params.snapId
         });
       } else {
-        store.dispatch("movement/fetchMembers", {
-          movId: route.params.movId,
+        store.dispatch('movement/fetchMembers', {
+          movId: route.params.movId
         });
       }
     });
@@ -358,7 +362,7 @@ export default {
       // get initial data from server (1st page)
       onRequest({
         pagination: pagination.value,
-        filter: filter.value,
+        filter: filter.value
       });
     });
 
@@ -376,21 +380,21 @@ export default {
       setFilterQuery,
       loading,
       isFullscreen,
-      onRequest,
+      onRequest
     };
   },
   components: {
     // 'mt-member-node':defineAsyncComponent(() => import('./../components/mt-member-node.vue'),
-    "mt-member-view-menu": defineAsyncComponent(() =>
-      import("./../components/actions/mt-member-view-menu.vue")
+    'mt-member-view-menu': defineAsyncComponent(() =>
+      import('./../components/actions/mt-member-view-menu.vue')
     ),
-    "mt-member-context-menu": defineAsyncComponent(() =>
-      import("./../components/actions/mt-member-context-menu.vue")
+    'mt-member-context-menu': defineAsyncComponent(() =>
+      import('./../components/actions/mt-member-context-menu.vue')
     ),
-    "mt-batch": defineAsyncComponent(() =>
-      import("./../components/mt-batch.vue")
-    ),
-  },
+    'mt-batch': defineAsyncComponent(() =>
+      import('./../components/mt-batch.vue')
+    )
+  }
 };
 </script>
 
