@@ -4,8 +4,10 @@
     <mt-filter-sort :tab="tab" />
     <q-scroll-area style="height: calc(100vh - 50px); max-width: 100%">
       <div
-        style="padding-left: 56px; min-height: calc(100vh - 50px)"
-        :style="`background: ${q.dark.isActive ? '#263238' : 'white'};`"
+        style="min-height: calc(100vh - 50px)"
+        :style="
+          `background: ${q.dark.isActive ? 'var(--q-dark-page)' : 'white'};`
+        "
         data-cy="background"
       >
         <MovBanner />
@@ -31,21 +33,39 @@
                 : 'img:icons/file-tree.svg'
             "
           >
+            <div
+              class="q-mx-sm text-bold"
+              :class="q.dark.isActive ? 'text-white' : 'text-black'"
+            >
+              -
+            </div>
             <q-select
               v-model="visibleTree"
               :options="treeOpts"
               :color="q.dark.isActive ? 'white' : 'black'"
               borderless
               dense
-              class="q-ml-sm"
               map-options
               hide-dropdown-icon
               option-value="id"
             >
               <template v-slot:selected>
                 <div :style="`color:${color}`">
-                  {{ visibleTree ? visibleTree.label : '' }}
+                  {{
+                    visibleTree && visibleTree.label > ''
+                      ? visibleTree.label
+                      : 'Untitled Tree'
+                  }}
                 </div>
+              </template>
+              <template v-slot:option="scope">
+                <q-item v-bind="scope.itemProps">
+                  <q-item-section>
+                    <q-item-label>{{
+                      scope.opt.label > '' ? scope.opt.label : 'Untitled Tree'
+                    }}</q-item-label>
+                  </q-item-section>
+                </q-item>
               </template>
             </q-select>
           </q-tab>
@@ -54,8 +74,14 @@
 
         <q-separator />
 
-        <q-tab-panels v-model="tab" animated>
-          <q-tab-panel name="trees" class="q-px-none">
+        <q-tab-panels v-model="tab" animated class="q-mx-sm">
+          <q-tab-panel
+            name="trees"
+            class="q-px-none"
+            :style="
+              `background: ${q.dark.isActive ? 'var(--q-dark-page)' : 'white'};`
+            "
+          >
             <MovTrees
               :treeOpt="visibleTree"
               v-if="visibleTree !== null"
@@ -83,25 +109,41 @@
       :totalStats="totalStats"
       :complexStats="complexStats"
       :calcStats="calcStats"
+      :treeOpt="visibleTree"
     />
     <mt-legend-drawer :roleOpts="roleOpts" :modOpts="modOpts" />
   </div>
 </template>
 
-<script>
-import { defineAsyncComponent, ref, computed, watch } from 'vue';
+<script lang="ts">
+import {
+  defineAsyncComponent,
+  ref,
+  computed,
+  watch,
+  defineComponent
+} from 'vue';
 import { useQuasar } from 'quasar';
 import { useStore } from 'vuex';
 import { useRoute } from 'vue-router';
 
-export default {
+export default defineComponent({
   name: 'MovementPage',
   setup() {
+    interface Tree {
+      id: string;
+      label: string;
+      tree: [];
+    }
     const q = useQuasar();
     const store = useStore();
     const route = useRoute();
     const tab = ref('trees');
-    const visibleTree = ref(null);
+    const visibleTree = ref<null | Tree>(null);
+    if (store.state.movement.movement.defaultTree) {
+      visibleTree.value =
+        store.state.movement.trees[store.state.movement.movement.defaultTree];
+    }
     const mods = computed(() => store.state.movement.mods);
     const roles = computed(() => store.state.movement.roles);
     const trees = computed(() => store.state.movement.trees);
@@ -124,11 +166,11 @@ export default {
       q.localStorage.has('sortKey') ? q.localStorage.getItem('sortKey') : 'Name'
     );
 
-    function fetchTreeData(e) {
+    function fetchTreeData(e: { movId: string; treeId: string | null }) {
       return store.dispatch('movement/fetchTreeData', e);
     }
 
-    function setCurrentTree(e) {
+    function setCurrentTree(e: Tree | null) {
       return store.commit('movement/setCurrentTree', e);
     }
     watch(
@@ -140,8 +182,10 @@ export default {
         ) {
           setCurrentTree(visibleTree.value);
           fetchTreeData({
-            movId: route.params.movId,
-            treeId: visibleTree.value.id,
+            movId: Array.isArray(route.params.movId)
+              ? route.params.movId[0]
+              : route.params.movId,
+            treeId: visibleTree.value ? visibleTree.value.id : null
           });
         }
       },
@@ -154,22 +198,23 @@ export default {
         // treeSelect.__updateMenu(true);
         if (
           (visibleTree.value === null && newVal > [] && newVal !== oldVal) ||
-          !newVal.find((val) => {
-            return visibleTree.value.id === val.id;
+          !newVal.find((val: Tree) => {
+            return visibleTree.value && visibleTree.value.id === val.id;
           })
         ) {
           visibleTree.value = { ...newVal[0] };
         } else if (
-          newVal > [] &&
-          visibleTree.value.label !==
-            newVal.find((val) => {
-              return visibleTree.value.id === val.id;
-            }).label
+          newVal > [] && visibleTree.value
+            ? visibleTree.value.label
+            : null !==
+              newVal.find((val: Tree) => {
+                return visibleTree.value && visibleTree.value.id === val.id;
+              }).label
         ) {
           visibleTree.value = {
-            ...newVal.find((val) => {
-              return visibleTree.value.id === val.id;
-            }),
+            ...newVal.find((val: Tree) => {
+              return visibleTree.value && visibleTree.value.id === val.id;
+            })
           };
         }
       },
@@ -194,19 +239,19 @@ export default {
       treeOpts,
       treeSorted,
       trees,
-      members,
+      members
     };
   },
 
   preFetch({ store, currentRoute }) {
     store.dispatch('movement/fetchTrees', {
-      movId: currentRoute.params.movId,
+      movId: currentRoute.params.movId
     });
     store.dispatch('movement/fetchMemberList', {
-      movId: currentRoute.params.movId,
+      movId: currentRoute.params.movId
     });
     store.dispatch('movement/fetchStyles', {
-      movId: currentRoute.params.movId,
+      movId: currentRoute.params.movId
     });
   },
   components: {
@@ -233,7 +278,7 @@ export default {
     ),
     'mt-filter-sort': defineAsyncComponent(() =>
       import('./../components/mt-filter-sort.vue')
-    ),
-  },
-};
+    )
+  }
+});
 </script>
