@@ -1,18 +1,20 @@
 import { getDatabase, ref, get, query, orderByChild, startAt, endAt } from "@firebase/database";
 
 export function fetchTrend({commit, state}, {movId, styleId, treeId, startDate, endDate}) {
-  // console.log(`/movements/${movId}/trends/${styleId}/${treeId}`, new Date(endDate), new Date(1640995200000))
+  // console.log(`/movements/${movId}/trends/${styleId}/${treeId}`, startDate, endDate)
   return get(query(
       ref(
         getDatabase(),
         `/movements/${movId}/trends/${styleId}/${treeId}`
-      ), orderByChild('date'),startAt(startDate), endAt(endDate))).then(snapshot => {
+      ), orderByChild('date'), startAt(startDate), endAt(endDate))).then(snapshot => {
+        // console.log(snapshot.val())
+        if(!snapshot.val()) return {}
         let trend = []
         for(let ii in snapshot.val()) {
           trend.push(snapshot.val()[ii])
         }
         // console.log(trend)
-        commit('setTrend', { movId, styleId, treeId, trend: trend, computed: {mean: mean(trend), median: median(trend), mode: mode(trend), min: min(trend), max: max(trend) }})
+        commit('setTrend', { movId, styleId, treeId, trend: trend.sort((a, b) => {return a.date - b.date}), computed: {mean: mean(trend), median: median(trend), mode: mode(trend), min: min(trend), max: max(trend) }})
         return { movId, styleId, treeId, trend: trend, computed: {mean: mean(trend), median: median(trend), mode: mode(trend), min: min(trend), max: max(trend) }}
       })
 }
