@@ -252,7 +252,7 @@ describe('access page tests', () => {
         cy.get('[data-cy="name"]', {timeout: 10000}).should('contain', 'Anonymous')
         // cy.get('[data-cy="avatar"]', {timeout: 10000}).should('have.attr', 'src')
         cy.get('[data-cy="email"]', {timeout: 10000}).should('contain', 'test@email.com')
-        cy.get('[data-cy="permission"]', {timeout: 10000}).should('contain', 'owner')
+        cy.get('[data-cy="permission"]', {timeout: 10000}).should('contain', 'Owner')
         cy.get('[data-cy="actions"]', {timeout: 10000}).within(()=>{
         // cy.get('[data-cy="delete"]', {timeout: 10000}).should('contain', 'delete')
         })
@@ -275,7 +275,7 @@ describe('access page tests', () => {
           cy.get('[data-cy="name"]', {timeout: 10000}).should('contain', 'Test User')
           // cy.get('[data-cy="avatar"]', {timeout: 10000}).should('have.attr', 'src')
           cy.get('[data-cy="email"]', {timeout: 10000}).should('contain', 'test@user.com.au')
-          cy.get('[data-cy="permission"]', {timeout: 10000}).should('contain', 'viewer')
+          cy.get('[data-cy="permission"]', {timeout: 10000}).should('contain', 'Viewer')
           cy.get('[data-cy="actions"]', {timeout: 10000}).within(()=>{
           cy.get('[data-cy="delete"]', {timeout: 10000}).should('contain', 'delete').click()
           })
@@ -311,7 +311,7 @@ describe('access page tests', () => {
         })
 
         cy.root().get('[data-cy="user-test@user.com.au"]', {timeout: 10000}).within(() => {
-          cy.dataCy('permission').should('contain', 'editor')
+          cy.dataCy('permission').should('contain', 'Editor')
         })
     });
 
@@ -414,12 +414,13 @@ describe('access page tests', () => {
       })
       cy.dataCy('access-checkboxes').within(() => {
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
         cy.dataCy('invite').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('remove').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('invite').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('remove').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
@@ -443,11 +444,12 @@ describe('access page tests', () => {
       })
       cy.dataCy('member-checkboxes').within(() => {
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('export').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
@@ -495,40 +497,44 @@ describe('access page tests', () => {
       })
         cy.dataCy('roles').within(() => {
           cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         })
         cy.dataCy('mods').within(() => {
           cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         })
         cy.dataCy('complex').within(() => {
           cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         })
         cy.dataCy('calc').within(() => {
           cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+          cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('edit').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
           cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
@@ -548,8 +554,9 @@ describe('access page tests', () => {
       })
       cy.dataCy('snapshots-checkboxes').within(() => {
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
-        cy.dataCy('update').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+        cy.dataCy('update').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('update').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
       })
     });
@@ -571,12 +578,13 @@ describe('access page tests', () => {
       })
       cy.dataCy('tree-checkboxes').within(() => {
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('export').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('import').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('remove-import').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('create').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('export').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('delete').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
@@ -600,10 +608,11 @@ describe('access page tests', () => {
       })
       cy.dataCy('trends-checkboxes').within(() => {
         cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')}).click()
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
         cy.dataCy('view-mg').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('create-mg').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
         cy.dataCy('delete-mg').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')}).click()
-        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--falsy').should('exist')})
+        cy.dataCy('view').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('view-mg').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('create-mg').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
         cy.dataCy('delete-mg').within(()=>{cy.get('.q-checkbox__inner--truthy').should('exist')})
